Implement revenue and expense display in Account Fragment

+++ app/src/main/java/com/example/bookkeeper/utils/ListAdapter.kt
@@ -1,7 +1,8 @@
 package com.example.bookkeeper.utils
 
+import android.content.Context
+import android.content.res.Resources
 import android.graphics.Color
-import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
@@ -11,6 +12,7 @@ import androidx.recyclerview.widget.RecyclerView
 import com.example.bookkeeper.R
 import com.example.bookkeeper.model.Entry
 import com.example.bookkeeper.model.EntryVM
+import com.google.android.material.dialog.MaterialAlertDialogBuilder
 import kotlinx.android.synthetic.main.list_item.view.*
 
 //class ListAdapter(
@@ -18,14 +20,18 @@ import kotlinx.android.synthetic.main.list_item.view.*
 //    private val listener: EntryListItemClickNotify, private val entryVM: EntryVM
 //) :
 class ListAdapter(
-    private val entryVM: EntryVM
+    private val entryVM: EntryVM,
 ) :
     RecyclerView.Adapter<ListAdapter.MyViewHolder>() {
     private var entryList = emptyList<Entry>()
     private lateinit var moneyFormatHelper: MoneyFormatHelper
+    private lateinit var context: Context
+    private lateinit var resources: Resources
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): MyViewHolder {
-        moneyFormatHelper = MoneyFormatHelper(parent.context)
+        context = parent.context
+        resources = context.resources
+        moneyFormatHelper = MoneyFormatHelper(context)
         return MyViewHolder(
             LayoutInflater.from(parent.context).inflate(R.layout.list_item, parent, false)
         )
@@ -49,7 +55,6 @@ class ListAdapter(
 
     override fun onBindViewHolder(holder: MyViewHolder, position: Int) {
         val currentItem = entryList[position]
-        Log.d("[TEST]", currentItem.date.toString())
         holder.itemView.descriptionTV.text = currentItem.description
         holder.itemView.amountTV.text = moneyFormatHelper.getPrefMoneyString(currentItem.amount!!)
         if (currentItem.amount > 0) {
@@ -63,8 +68,14 @@ class ListAdapter(
                 .toString().substring(0, 5)
         holder.itemView.timeTV.text = string
         holder.delImgBtn.setOnClickListener {
-            Log.v("Img Button", "Clicked $currentItem")
-            entryVM.deleteEntry(currentItem)
+            MaterialAlertDialogBuilder(context)
+                .setTitle(resources.getString(R.string.dialog_delete_title))
+                .setNegativeButton(resources.getString(R.string.dialog_delete_cancel)) { _, _ ->
+                }
+                .setPositiveButton(resources.getString(R.string.dialog_delete_confirm)) { _, _ ->
+                    entryVM.deleteEntry(currentItem)
+                }
+                .show()
         }
         holder.viewItem.setOnClickListener {
             //TODO: implement click to modify.
